Refactoring hmmcalibrate, in progress. (not a working rev.)
git-svn-id: https://svn.janelia.org/eddylab/eddys/src/hmmer/trunk@1387 db3e70e4-12e1-0310-b23f-81a6e3d09ba7

+++ src/histogram.c
@@ -669,7 +669,7 @@ FITFAILED:
  *                        else 2 if mu, lambda are estimated from data
  */
 void
-ExtremeValueSetHistogram(struct histogram_s *h, float mu, float lambda, 
+ExtremeValueSetHistogram(struct histogram_s *h, double mu, double lambda, 
 			 float lowbound, float highbound, int ndegrees)
 {
   int   sc;
